api endpoint for posting days, with concerns for both day controllers
